requirements.txt
fastapi
uvicorn
requests
tomli
uuid
pymongo
pydantic
typing
python-dotenv
aiohttp
aiofiles
cloudinary
pystyle
matplotlib
seaborn
typeguard
psutil
generateApiKey
slowapi
hashlib
httpx